refactored classes, created game_display module
