+++ runner.rb
@@ -4,6 +4,7 @@ require './lib/placement_validation'
 require './lib/board'
 require './lib/computer'
 require './lib/player'
+require './lib/game_display'
 require './lib/game'
 
 game = Game.new
